+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [CU-mdwa93] Short content product in Item by [lamah.dev@gmail.com].
 - [CU-e5k8uy] Report for invitation users by [kienluu612@gmail.com] [minhnguyenit14@gmail.com].
 - [CU-kbhzen] Open store from gpsListStore by open_store_from_list_key by [minhnguyenit14@gmail.com].
+- [CU-m2fhv6] POSSection in Confirm by [kienluu612@gmail.com].
 - [CU-m2hhc0] Use Linking to open any url not starting with "http" by [minhnguyenit14@gmail.com].
 - [CU-kben7m] not_allow_edit_cart_key for preventing editing order by [kienluu612@gmail.com].
 ### Changed